requirements.txt
openai
google-cloud-speech
fpdf
flask
flask-cors